code

+++ src/warenautomat/Ware.java
@@ -4,10 +4,10 @@ import java.util.Date;
 public class Ware {
 
 	private String name;
-	private double price;
+	private int price;
 	private Date date;
 
-	public Ware(String name, double price, Date date) {
+	public Ware(String name, int price, Date date) {
 		this.name = name;
 		this.price = price;
 		this.date = date;
@@ -18,7 +18,7 @@ public class Ware {
 		return name;
 	}
 
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
